cleaner stricter type on lorodoc for sync plugin
