+++ src/components/editor/types.ts
@@ -1,0 +1,1 @@
+export type { LoroDocType as LoroDocTypeCustom } from "loro-prosemirror"
